added http client for basic endpoints
